requirements.txt
annotated-types==0.7.0
black==24.8.0
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
greenlet==3.0.3
idna==3.8
jmespath==1.0.1
JPype1==1.5.0
mypy-extensions==1.0.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pathspec==0.12.1
platformdirs==4.3.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.32
tabula-py==2.9.3
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
